+++ lib/Command/ListDuplicates.php
@@ -2,26 +2,9 @@
 namespace OCA\DuplicateFinder\Command;
 
 use OC\Core\Command\Base;
-use OC\Files\Search\SearchQuery;
-use OC\Files\Search\SearchComparison;
-use OC\Files\Search\SearchOrder;
-use OC\Files\Utils\Scanner;
 use OCP\Encryption\IManager;
-use OCP\Files\File;
-use OCP\Files\Folder;
-use OCP\Files\NotFoundException;
-use OCP\Files\Search\ISearchComparison;
-use OCP\EventDispatcher\IEventDispatcher;
-use OCP\IConfig;
 use OCP\IDBConnection;
-use OCP\IPreview;
-use OCP\IUser;
 use OCP\IUserManager;
-use OCP\AppFramework\Http\DataResponse;
-use OCA\Files\Helper;
-use OC\Files\Filesystem;
-use Exception;
-use Symfony\Component\Console\Input\InputArgument;
 use Symfony\Component\Console\Input\InputInterface;
 use Symfony\Component\Console\Input\InputOption;
 use Symfony\Component\Console\Output\OutputInterface;
@@ -88,7 +71,7 @@ class ListDuplicates extends Base
             return 1;
         }
         $user = $input->getOption('user');
-
+        $result = 0;
         if ($user) {
             if ($user === true) {
                 $this->output->writeln('User parameter has an invalid value.');
@@ -98,27 +81,41 @@ class ListDuplicates extends Base
             } else {
                 $users = $user;
             }
-            foreach ($users as $user) {
-                if (!$this->userManager->userExists($user)) {
-                    $this->output->writeln('User '.$user.' is unkown.');
-                    return 1;
-                }
-                CMDUtils::showDuplicates(
-                    $this->fileDuplicateService,
-                    $this->fileInfoService,
-                    $this->output,
-                    function () {
-                        $this->abortIfInterrupted();
-                    },
-                    $user
-                );
+            if ($result === 0) {
+                $result = $this->listDuplicatesForUsers($users);
             }
         } else {
-            CMDUtils::showDuplicates($this->fileDuplicateService, $this->fileInfoService, $this->output, function () {
+            CMDUtils::showDuplicates($this->fileDuplicateService, $this->output, function () {
                 $this->abortIfInterrupted();
             });
+            $result = 0;
         }
 
-        return 0;
+        return $result;
+    }
+
+    /**
+     * @param array<string> $users
+     */
+    private function listDuplicatesForUsers(array $users) : int
+    {
+        $result = 0;
+        foreach ($users as $user) {
+            if (!$this->userManager->userExists($user)) {
+                $this->output->writeln('User '.$user.' is unkown.');
+                $result = 1;
+                break;
+            }
+            CMDUtils::showDuplicates(
+                $this->fileDuplicateService,
+                $this->output,
+                function () {
+                    $this->abortIfInterrupted();
+                },
+                $user
+            );
+        }
+        unset($user);
+        return $result;
     }
 }