+++ lib/Event/NewFileInfoEvent.php
@@ -1,23 +1,6 @@
 <?php
 namespace OCA\DuplicateFinder\Event;
 
-use OCP\EventDispatcher\Event;
-use OCA\DuplicateFinder\Db\FileInfo;
-
-class NewFileInfoEvent extends Event
+class NewFileInfoEvent extends AbstractFileInfoEvent
 {
-
-  /** @var FileInfo */
-    private $fileInfo;
-
-    public function __construct(FileInfo $fileInfo)
-    {
-        parent::__construct();
-        $this->fileInfo = $fileInfo;
-    }
-
-    public function getFileInfo(): FileInfo
-    {
-        return $this->fileInfo;
-    }
 }